Backed out 5 changesets (bug 965022) for mochitest-2 failures on a CLOSED TREE.
Backed out changeset 7f68752ffe1e (bug 965022)
Backed out changeset 24251d4da019 (bug 965022)
Backed out changeset a34ae046c947 (bug 965022)
Backed out changeset f747fd154739 (bug 965022)
Backed out changeset f076faf3c282 (bug 965022)

UltraBlame original commit: 894b7372561dcf1a27cb4038a4f73c44cb16cbb2

+++ layout/style/nsStyleUtil.cpp
@@ -62,7 +62,7 @@ void nsStyleUtil::AppendEscapedCSSString(const nsAString& aString,
   for (; in != end; in++) {
     if (*in < 0x20 || (*in >= 0x7F && *in < 0xA0)) {
       
-      aReturn.AppendPrintf("\\%hX ", static_cast<int16_t>(*in));
+      aReturn.AppendPrintf("\\%hX ", *in);
     } else {
       if (*in == '"' || *in == '\'' || *in == '\\') {
         
@@ -116,7 +116,7 @@ nsStyleUtil::AppendEscapedCSSIdent(const nsAString& aIdent, nsAString& aReturn)
   
   
   if (in != end && ('0' <= *in && *in <= '9')) {
-    aReturn.AppendPrintf("\\%hX ", static_cast<int16_t>(*in));
+    aReturn.AppendPrintf("\\%hX ", *in);
     ++in;
   }
 
@@ -127,7 +127,7 @@ nsStyleUtil::AppendEscapedCSSIdent(const nsAString& aIdent, nsAString& aReturn)
     }
     if (ch < 0x20 || (0x7F <= ch && ch < 0xA0)) {
       
-      aReturn.AppendPrintf("\\%hX ", static_cast<int16_t>(*in));
+      aReturn.AppendPrintf("\\%hX ", *in);
     } else {
